% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{essay0_revised_and_shuffled}
\alias{essay0_revised_and_shuffled}
\title{Cleaned (shuffled) OkCupid essay data}
\format{
A \code{data.frame} with 59946 rows and 1 variable:
\describe{
  \item{essay0}{First 140 characters of first essay response (my self summary)}
}
}
\source{
\url{https://github.com/rudeboybert/JSE_OkCupid}
}
\usage{
essay0_revised_and_shuffled
}
\description{
Cleaned randomly row-shuffled essay data of 59,946 OkCupid users. To
unpair essays from users, the rows of the essay data were randomly
scrambled. Furthermore, due to R package size restrictions, only the
first 140 characters of each user's first essay response (my self
summary) is included. The original data and codebook can be found at
\url{https://github.com/rudeboybert/JSE_OkCupid}.
}
\examples{
library(tibble)
essay0_revised_and_shuffled

# If using RStudio:
# View(essay0_revised_and_shuffled)
}
\keyword{datasets}
