% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okcupiddata.R
\docType{package}
\name{okcupiddata}
\alias{okcupiddata}
\alias{okcupiddata-package}
\title{okcupiddata:}
\description{
Profile data from .
}
\section{Data sets}{

Type \code{data(package="okcupiddata")} to view all datasets.
}
\examples{
library(okcupiddata)
data(profiles)
data(profilesdata)
}
\references{
Albert Y. Kim, Adriana Escobedo-Land (2015). OkCupid Profile Data for
Introductory Statistics and Data Science Courses. Journal of Statistics
Education, 23(2). \url{http://www.amstat.org/publications/jse/v23n2/kim.pdf}.
}

