% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profiles}
\alias{profiles}
\title{(Cleaned) OkCupid profile data}
\format{A tibble (\code{tbl_df} class data frame in \link[dplyr]{dplyr} style) with 59946 rows and 22 variables:}
\source{
\url{https://github.com/rudeboybert/JSE_OkCupid}
}
\usage{
profiles
}
\description{
Cleaned version of \code{\link{profiles_raw}} meant for novices. Differences
between \code{profiles} and \code{profiles_raw} are described in Details.
}
\details{
Differences between cleaned and raw version of profiles data.
\describe{
\item{Missing \code{income} values}{Previously stored as \code{-1} is now stored as \code{NA}'s}
\item{All other missing values}{Previously stored as \code{""} is now stored as \code{NA}'s}
\item{\code{offspring} and \code{sign}}{String instances \code{"?&rsquo;"} are replaced with apostrophes.}
\item{\code{last_online}}{Converted to USA/Pacific timezone POSIXct date-time objects using \link[lubridate]{parse_date_time}}
}
}
\examples{
library(okcupiddata)
data(profiles)
# If using RStudio:
# View(profiles)
summary(profiles$income)
}
\keyword{datasets}

