% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profiles_raw}
\alias{profiles_raw}
\title{(Cleaned) OkCupid profile data}
\format{A tibble (\code{tbl_df} class data frame in \link[dplyr]{dplyr} style) with 59946 rows and 22 variables:
\describe{
  \item{ID}{Anonymized ID number of user. Used to join with essay data}
  \item{age}{Age}
  \item{body_type}{Body type}
  \item{diet}{Dietary habits}
  \item{drinks}{Drinking habits}
  \item{drugs}{Drug usage habits}
  \item{education}{Education level}
  \item{ethnicity}{Ethnicity}
  \item{height}{Height in inches}
  \item{income}{Income}
  \item{job}{Job}
  \item{last_online}{Date/time of last login to OkCupid}
  \item{location}{Location}
  \item{offspring}{Number of offspring}
  \item{orientation}{Sexual orientation}
  \item{pets}{Number of pets}
  \item{religion}{Religious affiliation}
  \item{sex}{Sex. Note at the time OkCupid only allowed for male/female binary. This has since been relaxed.}
  \item{sign}{Astrological sign}
  \item{smokes}{Smoking habits}
  \item{speaks}{Languages spoken}
  \item{status}{Relationship status}
}}
\source{
\url{https://github.com/rudeboybert/JSE_OkCupid}
}
\usage{
profiles_raw
}
\description{
Raw profile data of 59,946 OkCupid users who were living within 25 miles of San
Francisco, had active profiles on June 26, 2012, were online in the previous
year, and had at least one picture in their profile. Essay response data were
not included. Full codebook can be found
at \url{https://github.com/rudeboybert/JSE_OkCupid}
}
\examples{
library(okcupiddata)
data(profiles_raw)
# If using RStudio:
# View(profiles_raw)
summary(profiles_raw$income)
}
\keyword{datasets}

