% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profiles_raw}
\alias{profiles_raw}
\title{Prices of 50,000 round cut diamonds.}
\format{A data frame with 59946 rows and 22 variables:
\describe{
  \item{ID}{Anonymized ID number of user. Used to join with essay data}
  \item{age}{age}
  \item{body_type}{body type}
  \item{diet}{diet}
  \item{drinks}{drinks}
  \item{drugs}{drugs}
  \item{education}{drinks}
  \item{ethnicity}{drinks}
  \item{height}{drinks}
  \item{income}{drinks}
  \item{job}{drinks}
  \item{last_online}{drinks}
  \item{location}{drinks}
  \item{offspring}{drinks}
  \item{orientation}{drinks}
  \item{pets}{drinks}
  \item{religion}{drinks}
  \item{sex}{drinks}
  \item{sign}{drinks}
  \item{smokes}{drinks}
  \item{speaks}{drinks}
  \item{status}{drinks}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
profiles_raw
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

